+++ src/cone/ugm/browser/group.py
@@ -9,39 +9,35 @@ from cone.tile import tile
 from cone.ugm.browser import form_field_definitions
 from cone.ugm.browser.authoring import AddFormFiddle
 from cone.ugm.browser.authoring import EditFormFiddle
-from cone.ugm.browser.columns import Column
 from cone.ugm.browser.listing import ColumnListing
 from cone.ugm.browser.principal import PrincipalForm
 from cone.ugm.browser.roles import PrincipalRolesForm
 from cone.ugm.model.group import Group
 from cone.ugm.model.utils import ugm_general
 from cone.ugm.model.utils import ugm_groups
-from plumber import plumber
+from plumber import plumbing
 from pyramid.i18n import TranslationStringFactory
 from pyramid.security import has_permission
 from webob.exc import HTTPFound
 from yafowil.base import ExtractionError
 from yafowil.utils import UNSET
+import urllib2
 
 
 _ = TranslationStringFactory('cone.ugm')
 
 
-@tile('leftcolumn', interface=Group, permission='view')
-class GroupLeftColumn(Column):
-
-    add_label = _('add_group', default='Add Group')
+@tile('leftcolumn', 'templates/principal_left_column.pt',
+      interface=Group, permission='view')
+class GroupLeftColumn(Tile):
 
     @property
-    def can_add(self):
-        return has_permission('add_group', self.model.parent, self.request)
-
-    def render(self):
-        setattr(self.request, '_curr_listing_id', self.model.name)
-        return self._render(self.model.parent, 'leftcolumn')
+    def principals_target(self):
+        query = make_query(pid=self.model.name)
+        return make_url(self.request, node=self.model.parent, query=query)
 
 
-@tile('rightcolumn', 'templates/right_column.pt',
+@tile('rightcolumn', 'templates/principal_right_column.pt',
       interface=Group, permission='view')
 class GroupRightColumn(Tile):
 
@@ -147,8 +143,6 @@ class UsersOfGroupColumnListing(ColumnListing):
     css = 'users'
     slot = 'rightlisting'
     list_columns = ColumnListing.user_list_columns
-    # XXX: Why items and not keys() / __iter__?
-    # used to be a readonly property
     query_items = Principals(members_only=True)
     batchname = 'rightbatch'
 
@@ -159,8 +153,6 @@ class AllUsersColumnListing(ColumnListing):
     css = 'users'
     slot = 'rightlisting'
     list_columns = ColumnListing.user_list_columns
-    # XXX: Why items and not keys() / __iter__?
-    # used to be a readonly property
     query_items = Principals()
     batchname = 'rightbatch'
 
@@ -174,6 +166,7 @@ class AllUsersColumnListing(ColumnListing):
 class InOutListing(ColumnListing):
     selected_items = Principals(members_only=True)
     available_items = Principals(available_only=True)
+    display_control_buttons = True
 
     @property
     def user_attrs(self):
@@ -184,12 +177,6 @@ class InOutListing(ColumnListing):
     def user_default_sort_column(self):
         settings = ugm_general(self.model)
         return settings.attrs['user_display_name_attr']
-
-    @property
-    def display_control_buttons(self):
-        return True
-        #settings = ugm_general(self.model)
-        #return settings.attrs['controls_membership_assignment_widget']
 
 
 class GroupForm(PrincipalForm):
@@ -217,10 +204,8 @@ class GroupForm(PrincipalForm):
 
 
 @tile('addform', interface=Group, permission="add_group")
+@plumbing(AddBehavior, PrincipalRolesForm, AddFormFiddle)
 class GroupAddForm(GroupForm, Form):
-    __metaclass__ = plumber
-    __plumbing__ = AddBehavior, PrincipalRolesForm, AddFormFiddle
-
     show_heading = False
     show_contextmenu = False
 
@@ -248,11 +233,10 @@ class GroupAddForm(GroupForm, Form):
     def next(self, request):
         next_resource = self.request.environ.get('next_resource')
         if next_resource:
-            url = make_url(request.request,
-                           node=self.model,
-                           resource=next_resource)
+            query = make_query(pid=next_resource)
+            url = make_url(request.request, node=self.model.parent, query=query)
         else:
-            url = make_url(request.request, node=self.model)
+            url = make_url(request.request, node=self.model.parent)
         if self.ajax_request:
             return [
                 AjaxAction(url, 'leftcolumn', 'replace', '.left_column'),
@@ -262,10 +246,8 @@ class GroupAddForm(GroupForm, Form):
 
 
 @tile('editform', interface=Group, permission="edit_group", strict=False)
+@plumbing(EditBehavior, PrincipalRolesForm, EditFormFiddle)
 class GroupEditForm(GroupForm, Form):
-    __metaclass__ = plumber
-    __plumbing__ = EditBehavior, PrincipalRolesForm, EditFormFiddle
-
     show_heading = False
     show_contextmenu = False
 
@@ -284,7 +266,12 @@ class GroupEditForm(GroupForm, Form):
         self.model.model.context()
 
     def next(self, request):
-        url = make_url(request.request, node=self.model)
+        came_from = request.get('came_from')
+        if came_from:
+            came_from = urllib2.unquote(came_from)
+            url = '{}?pid={}'.format(came_from, self.model.name)
+        else:
+            url = make_url(request.request, node=self.model)
         if self.ajax_request:
             return [
                 AjaxAction(url, 'leftcolumn', 'replace', '.left_column'),